Extract calibration file format dispatch into engine/src/depth/calibrationFileParser.ts
parseCalibrationFile was already fully pure - dispatches to
CalibTxtParser/YamlCalibrationParser/ColmapParser/ZedParser/
RealSenseParser/TumParser by file extension and content sniffing, no
`this` dependency at all. Moved as-is and dropped the six now-unused
parser-class imports from main.ts.

main.ts: 11,547 -> 11,437 lines.

Verified: engine tsc clean, full webpack compile clean, all 59
extension tests passing under xvfb-run.
